+++ source/include/ac_C_lib.h
@@ -124,7 +124,7 @@ typedef void (*AC_RETRYPTR)(void);
  * insufficient for the target bezdata, an error will be returned and *length
  * will be set to the desired size.
  */
-ACLIB_API int AutoColorString(const char *srcbezdata, const char *fontinfo, char *dstbezdata, int *length, int allowEdit, int allowHintSub, int roundCoords, int debug);
+ACLIB_API int AutoColorString(const char *srcbezdata, const char *fontinfo, char *dstbezdata, size_t *length, int allowEdit, int allowHintSub, int roundCoords, int debug);
 
 /*
  * Function: AC_initCallGlobals
